WIP respond to PR feedback
smartcitiesdata/smartcitiesdata#722
