+++ apps/andi/assets/js/app.js
@@ -29,20 +29,6 @@ Hooks.showSnackbar = {
     }
 }
 
-Hooks.Unload = {
-    unSavedModalDisplayed() {
-        return this.el.dataset.showUnsavedChangesModal == "true";
-    },
-    mounted() {
-        window.addEventListener("beforeunload", e => {
-            if (!this.unSavedModalDisplayed()) {
-                e.preventDefault();
-                e.returnValue = '';
-            }
-        })
-    }
-}
-
 let csrfToken = document.querySelector("meta[name='csrf-token']").getAttribute("content");
 let liveSocket = new LiveSocket('/live', Socket, {hooks: Hooks, params: {_csrf_token: csrfToken}})
 liveSocket.connect()
