+++ apps/andi/lib/andi/input_schemas/datasets/technical.ex
@@ -29,7 +29,6 @@ defmodule Andi.InputSchemas.Datasets.Technical do
     field(:sourceUrl, :string)
     field(:topLevelSelector, :string)
     field(:cadence, :string)
-
     field(:orgId, :string)
 
     belongs_to(:dataset, Dataset, type: :string, foreign_key: :dataset_id)
